Organizer event participant selection complete

+++ oblongApp/app/src/main/java/com/example/oblong/Event.java
@@ -1,10 +1,5 @@
 package com.example.oblong;
 
-import android.content.res.Resources;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-import android.graphics.ImageDecoder;
-import android.provider.MediaStore;
 import android.util.Log;
 
 import com.google.firebase.Timestamp;
@@ -15,7 +10,6 @@ import com.google.firebase.firestore.FirebaseFirestore;
 import java.io.Serializable;
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
-import java.util.Base64;
 import java.util.Date;
 import java.util.HashMap;
 
@@ -117,7 +111,7 @@ public class Event implements Serializable {
         this.eventCloseDate = eventCloseDate;
     }
 
-    public Long getEventCapacity() {
+    public int getEventCapacity() {
         return eventCapacity;
     }
